t2t-vit/requirements.txt
mindspore_gpu~=1.9.0
numpy~=1.21.6
Pillow~=9.4.0
PyYAML~=6.0